[OpenThread] Add SPI component, start async loop
This change adds SPI component to the OpenThread device. When an interrupt
is received from the radio, we go a SPI read of a few bytes and print it
out to compare against what's seen on the line.

Tests:
Driver unit test: ot_radio_test creates the device, issues a reset
                  and expects a valid header byte from the radio.
Manual test: Additional manual validation that the bytes printed out
             are the same as the ones seen on the logic analyzer trace.
Verified clean exit of the test by checking for [ot-radio] in `dm dump`.

Change-Id: I579d6260350faa64467aaaf2d6203c8a9c010d29