[build] Fix secret_spec rule with variants
The `secret_spec` rule generates its output with a toolchain-independent
root, which causes two separate rules outputing to the same location:
the default toolchain and the asan one. This commit fixes the rule to
make the non-default-toolchain version depend on the default one, ending
up with only one rule copying the file, which avoids ninja warnings. The
expected file is still output under out/default/secret_specs, where
infra will expect it to run the tests on CI, regardless of whether ASAN
is enabled or not.

Bug: 12348 # done
Test: fx set core.x64 --variant asan
  fx ninja -C out/default -w dupbuild=err -t query secret_specs/ciphertext/ledger_e2e_sync_credentials.ciphertext
  fx build
  ls -l out/default/secret_specs/ciphertext/
Change-Id: Ib7670c1ba0eeaa3e42158d9c2c58f0497c53bf9e