[streammap] Implement StreamMap
This CL implements the streammap lib, which is a stream of streams where
streams can be cancelled by key.

Some tasks serve many FIDL channels. With this library, N channels can
be unified in a single stream and any of those channels can be closed by
a key without waiting for new request.

TEST: fx run-test streammap_tests

Change-Id: I21c0a46225a6378e95eb7f35e6b9111a42648de3