[dart][l10n] Ermine shell localization example.
This change adds scaffolding that allows Ermine shell
to display some system messages in a language
other than English.

At the moment there is no way to select the locale of
the shell.  This will be added in followup
changes.  The resources are also not translated
since currently there is no infrastructure to do so.

Added the basic documentation that explains how the localization is
used and maintained.

Bug: 37241
Change-Id: I6414acf26f894564575fa4222cc386ebb93cd2a8