[audio_core] Replace downcasts with virtual function calls.
This CL trades fbl::RefPtr downcasts for virtual function calls in
AudioObject. While this CL moves functionality that may not
semantically "belong" inside AudioObject, this is simply a realization
of how existing code is interacting with these objects.

The intention is to centralize this functionality in AudioObject so
that we can continue to interate and improve on this abstraction.

Test: This is a refactor only; no functionality change.
      fx run-test audio_core_unittests
      fx run-test audio_fidl_tests
      fx run-test audio_device_tests
      fx run-test audio_service_tests
      fx run-test audio_pipeline_tests
Change-Id: I17c3df57a177035fb62a08ba07d98e7163f7b71f