[e2e] Kill Chromedriver if it fails to shut down.
This change makes the webdriver tearDown method
wait until it gets an exit code from chromedriver,
which signifies the process has ended.  The process
is passed a SIGKILL if it does not terminate in time.

This also causes webdriver logic to always find and
use unsued ports on the test host, to prevent clashes
with other processes.

Testing: host tests, run e2e test

Bug: 34882

Change-Id: I7c5144e999f273e36a629c6f410e5d11dbd23ccc