[kernel][x86] AMD: Configure LFENCE to be dispatch-serializing
LFENCE is an x86 instruction which can serve to constrain speculative
execution; it is a building block of most mitigations for certain
speculative execution information leaks (colloquially Spectre V1).

AMD CPUs have long supported LFENCE, but it is not dispatch-serializing
on Family 10h (and newer) CPUs unless a particular MSR bit is set.
Set that bit per vendor recommendation in the "Software Techniques
for Managing Speculation on AMD Processors" document, mitigation G-2.

Bug: 33667 Spectre mitigations?
Bug: 12540 Speculative Execution Mitigations.

Tested:
* Booted in QEMU + KVM on AMD Family 15h; checked MSR C0011029h:
$ k cpu rdmsr 0 0xc0011029
[00003.445] 01429:01431> CPU 0 RDMSR c0011029h val 2h

* Booted on AMD Family 17h (Ryzen 7 1700)
$ k cpu features
[141.756] 01412.01438> Vendor: AMD
[141.756] 01412.01438> Microarch: Zen
[141.756] 01412.01438> F/M/S: 17/1/1
[141.756] 01412.01438> patch_level: 8001137
[141.756] 01412.01438> Brand: AMD Ryzen 7 1700 Eight-Core Processor
[141.756] 01412.01438> Features: fpu sse sse2 sse3 ssse3 sse4.1 sse4.2 mmx avx avx2 fxsr xsave mon aesni clflush
[141.756] 01412.01438> clflushopt fsgsbase smep smap rdrand rdseed syscall nx huge rdtscp invar_tsc
[141.756] 01412.01438> Properties:
$ k cpu rdmsr 0 0xc0011029
[210.548] 01412.01438> CPU 0 RDMSR c0011029h val e2a002h

Change-Id: I1b7becf976ae9a9a60f8a6af90336372d6220e0b