[blackout] refactor host test library
This totally revamps the host-side test library for blackout. It adds
several new features - it will collect the common cli options
automatically, it has a concept of configurable test steps, breaking the
tests up into reusable chunks, and pulls the definitions of the common
steps into the library itself. After this refactor, host binary
definitions go from directly handling the gritty details of process
spawning and details about the contract between host and target tests to
being a few lines of straightforward configuration using a builder
pattern, while still providing the flexibility of performing arbitrary
operations.

It also bumps the sleep between writing bytes to the relay for a hard
reboot from 10 to 100 millis to make sure it caught the change (it must
be doing some kind of polling - it would miss the bit flip sometimes).

Change-Id: Ibb9c4502f63e4b18319e30199d2eac0eb9ddd617