[modular] Fix GN check errors in various modular targets
This fixes errors identified by 'gn check' when examining <bracket>
style includes in the modular codebase, mostly in tests. Several
targets were using libraries that they did not directly depend on. This
also refactors the modular config header targets to be a bit clearer.

This also expands the allowlist for the deprecated lib/svc/cpp to
reflect the current usage.

Bug: 2995, 36788
Change-Id: I20e9eb4dc340acba6afe2745ecbfaab534555d93