[build] Disable fvm-host-test
This change introduces the concept of `disabled` to the builtin zircon
test(), which if populated indicates that test is disabled, its value
being the reason why. When set on a host test, the consequence will be
the it is prevented from running.

In particular, this CL disables the fvm-host-test, which is taking 20+
minutes to run.

Bug: 37779
Change-Id: I3a1503c0f13b1af1da8bc066dd8c6d055f7b8dea