[camera] Removing fuchsia.camera.common
FIDL protocols are not allowed to have common in the name,
and rather than rename the fidl protocol, we are just
removing it, since it is deprecated and is a duplicate
of protocols in fuchsia.camera.

fuchsia.camera.test also depended on that library, and so
we're moving it to sdk to be with the other protocols.

Test: This involves renaming a few protocols and moving files,
so no functional change.

Change-Id: I0933f6806ccd63f22808d44abae5484cc80e1afc