[tools][debug] Check existence before uploading symbols
Currently, we don't check existence before uploading
a debug symbol. This is extremely wasteful as a symbol
may be present as part of many builds. Currently,
some invocations of the tool take more than 3 minutes,
where this change drops them to <10s.

Additionally fixes a bug where files are never closed.

In order to add unit tests for affected functionality,
this change adds a simple Bucket interface to split out
cloud storage calls.

This change was also tested by manually invoked tryjobs
which call this tool as a build step.

Bug: 37570

Change-Id: Iddf06fb1fc4c447f343897ffc9216fc5f72abccb