[bt][cli] Add optional filter for list-peers
When discovering devices, the list of peers can be very long and hard to
find the peer that you are searching for.  Add an optional argument to
the list-peers command to filter the peer list by identifier, address,
or peer name.

Update the output to indicate how many peers matched out of the total
number of known peers.

Test: Added unit test for testing match_peer function.
      Added unit test for get_peers function.
   - run bt-cli, start-discovery, list-peers with various input and observe output
Change-Id: Ic082ed4900269aacabde80209924468ce97c4199