[Settings] Use the Volume Control API for audio
Modified audio controller so that it uses the Volume Control
API to set the volume. This CL introduces the struct
StreamVolumeControl which contains the current Audio
Stream and a proxy for the Volume Control.

Bug: 37403
Testing:
There are no new behavior changes. There are changes to
audio_test.rs so that it’ll test that the volume is set correctly
through Volume Control instead of Audio Core
You can run the tests via:
   fx run-test setui_service_tests -- --test

Change-Id: I3af76bbec5d9473bf22f76986025cacb77449faf