[examples][fuzzer] Add FuzzedDataProvider example.
This CL adds an example of how to use the FuzzedDataProvider library
that comes with compiler-rt. The library helps in splitting the input
data to create fuzzers. Instead of making the programmer manually
consume data from the fuzzer input, the library provides helper methods
that can be used to shape the input data into basic data types.

Change-Id: I335442f9095be8a70cd2b683e5ecfecdda399fb6