[Settings] Add support for clearing accounts.
This changelist adds the ability for a setting service handler to
schedule clearing accounts. This is done via a flag in the device
settings service and invoked upon the next restart.

This changelist also calls this functionality from the system
controller to clear accounts when the login mode changes.

Fixes: fxb/37598
Test: fx run-test setui_service_tests -- --test
Testing Strategy: The system test has been modified to check
the device settings service to make sure the flag is set after
changing modes.

Change-Id: I6b3d2127df895c2fc9b1eb639b70bffc6038b459