[audio] Initialise SineSource's "frames_produced_" variable.
The SineSource doesn't initialise the "frames_produced_" variable.
Previously this was always 0 by luck, so tools like "audio tone" worked
as intended. However, we have recently started poisoning the stack with
0xAA bytes, tickling this bug and causing "audio tone" to immeadiately
exit.

We also add a basic test for SineSource that exercises its basic
functionality.

Sadly, tools like asan, msan, ubsan, or Clang's "--analyze"
options don't appear to detect this bug, so we are relying on
manual inspection to find other instances of it.

Change-Id: Id032bb3ee1d5245e347725b8b5c1990464ae62ca