[ledger] Remove children-manager-retainer cancels
My intention with these had been to avoid, during the destruction of a
ChildrenManager, signalling to Inspect that it should not call
detachers associated with that ChildrenManager, because those detachers
could then in turn empty the children, which could empty the
ChildrenManager, which could reënter the ChildrenManager's destructor.

I erred in not remembering that the semantics of the retainer's cancel
method are not "don't call the detachers, but do stop using the
retained ChildrenManager" but rather are "don't do anything" - so
calling the retainer's cancel method left Inspect with a dangling
ChildrenManager pointer (and if it held any detachers,  dangling
pointers to the children that created those detachers).

The calls are removed rather than tweaked because since c83d017822e6f54
emptying of children cannot synchronously empty the ChildrenManager
managing those children.

Change-Id: Idbc67ff345ae706e52022ebbc3bd0795dbfa26cd