[fdio][vx] Remove zx_vmo_replace_as_executable calls from fdio_spawn
This change is part of implementing the Verified Execution: Controls on
Executable Memory design. See go/fvx-executable-memory for more details.

This removes the major remaining case of fdio directly adding
ZX_RIGHT_EXECUTE to VMOs on the client side using the
zx_vmo_replace_as_executable syscall, and instead requests that the
filesystem server provide such a VMO if it is able to.

As of this change, users of fdio_spawn must either spawn new processes
from an executable that lives on a verified filesystem - either bootfs
or pkgfs - or add ZX_RIGHT_EXECUTE to the VMO to be spawned themselves
(and use fdio_spawn_vmo) if they are able to do so (soon JITs only).

This also adds an ugly hack in runtests-utils to unblock this change,
which will be properly fixed in fxb/37529.

Bug: 35238 #fixed
Bug: 37529

Change-Id: I1e3c2df94d0f6be96a648515734abfe74ef4f526