[audio_core] ChannelStrip class
Certain mix operations work most effectively when performed on one
channel at a time, but this is made more difficult if the data layout
is interleaved frame-by-frame. This CL introduces a simple class
(ChannelStrip) to store (and shift) audio data when transformed into
this layout. This class is not currently relied upon by any production
code but will be used by an incoming resampler.

Test: 4 new unittests in audio_mixer_unittests

Change-Id: I284fb8b788d487fb0379f139d6fdd2f18993a33e