[trace-engine] Account for null byte at end of string literals
If a trace event passes a string literal as an argument, there is a null
byte at the end which would get included in the inline string record.
This reduces the length by 1 in these cases which prevents the null byte
from appearing, but could cause an unexpected shortening if the argument
is meant as a byte array (for example).

Bug: 36404

Change-Id: If42b07f548d1b207b1412e592a32ec27dcfe589c