[media.sessions] Emit states instead of Deltas
After this CL, information emitted from the session service to clients
will be states, not deltas, so clients do not to keep track of state
themselves and can render their UI reactively.

This is valid according to the API contract, but not yet guaranteed.

To implement this, Stream is implemented for the Player proxy, which
emits a stream of its own states.

This CL also changes when clients see a new player's publication.
Previously clients might see it immediately when it is assigned an id.
Now, clients only see it once the player actually emits its status.

TEST: Tests added for Stream impl;
      fx run-test mediasession_tests
Change-Id: Ic5146aea2cf659310c51f506c5944a8a55583256