[bt][api] Publish the fuchsia.bluetooth.control library in the SDK
The fuchsia.bluetooth.control library includes the Control protocol for
interacting with the GAP layer of Bluetooth, which includes support for
the procedures required to connect and pair Bluetooth devices.

As it is necessary for partner code to perform these procedures, this
api needs to be exposed in the public SDK.

Bug: 37290

testability: No behavior is changed; only the access to the APIs.
Change-Id: Ibc21880f056af99209ec75d3239db54f416b1aff