[blackout] add separate setup step
The setup step is for doing test setup, such as formatting the disk.
Tests would occasionally take longer than the allotted test duration on
formatting the disk, which causes the test to cut the power during
formatting, which corrupts the disk, but not in the way we want. The
separate setup step moves that task outside of the timed test step, so
it can take as long as necessary.

Change-Id: I0163a711898330a53be648f9ed37d78fb3f6ce5b