[blackout] add retry loop to verification step
the loop attempts to ssh in 10 times, with a 1 second sleep between each
attempt. ssh returns an exit code of 255 if it fails itself, as opposed
to the command it's running failing, so if it's not that error code an
error is returned immediately. the number of retries and the duration of
the sleep between tries is configurable by individual tests.

Bug: 34504
Change-Id: I95e3e486ad503177427267e6b8a8f8c1c465d788