[host-target-testing][system-ota-tests] Paver should use pave-zedboot.sh
Before this patch, the system ota test would fail when the zedboot version
number changed, because there's no guarantee that the device has a
compatible zedboot installed on it. This patch switches to using the new
`pave-zedboot.sh` script, which first installs candidate build's zedboot
onto the device and boots into it, before trying to pave the rest of the
images onto the device.

Test: system-ota-tests, tested with a zedboot and minfs changing CL
Bug: 37872
Change-Id: I387e7e0d70ad07492913d834fa842c424508af24