[blackout] clean up error handling
Instead of using the provided Error struct from failure, we move to
having our own structured errors. This cleans up checking for command
success, the retry loop in verify, and consolidates printing out the
stdout and stderr when we don't like the result of a command on the
target. Most importantly, it enables a test runner to differentiate
between actual test failures and other types of errors, which will allow
for them to run the test over and over again and tabulate error types
over time with higher fidelity.

Change-Id: I0e821c2ffc704fc3c0f34b27f40a42eb11374b7e