[netsvc] Change default value of netsvc.disable
By default, netsvc.disable takes the value true (i.e. will be disabled).
This value is used by devcoordinator to determine whether to start
netsvc.

No change in behavior is expected, as it's explicitly not-disabled in
all configurations (in zedboot, and in the bringup product which is
inherited by all other product configurations).

Internal integration change 149553 has a run with green trybots for
reference.

Bug: 37043
Change-Id: Ia5b164e0e37e3d115a402ba0f2062117084fb33a