[fidlc] Behind a flag treat unions as non-simple
This is to make it easy to see where we'll need to update FIDL libraries
to remove the [Layout="Simple"] attribute once unions are out of line.

Turn this on by setting the gn flag:
zircon_extra_args={fidlc_union_not_simple=true}

This CL establishes an approach to selectively tweaking the behavior of
fidlc at build time. I plan to use this same approach for other
incomplete parts of our wire format transition.

Change-Id: I2b6e6a70d949686cdf38f178c588ff6d5d7c2e78