[zxtest] Avoid accidentally triggering copy constructors.
Specifically, in the context of (ASSERT|EXPECT)_(NULL|NOT_NULL)...  After being
handed of through a sequence of macros and lambdas, we end up in a context where
we are effectively saying something to the effect of the following (please note
that this is simplified just a bit)

```
bool op(const Foo& a, const Foo& b) { ... }

bool func(const auto& a, const auto& b) {
  return op(static_cast<Foo>(a), static_cast<Foo>(b));
}
```

When Foo is captured by the macro as a pointer-like type (for example,
std::unique_ptr or fbl::RefPtr), we run into trouble.  Even though op is
expecting a const Foo&, and even though func received a const Foo&, the static
cast to Foo ends up tripping implicit conversion via construction; in the cause
using the copy constructor.  Can I make a fbl::RefPtr<> from a const
fbl::RefPtr<>&?  Sure!  Just use the copy constructor.

In the case of std::unique_ptr, the copy constructor is deleted, so the macro
ends up failing to expand and users are forced to say something like

```
  ASSERT_NOT_NULL(my_uptr.get());
```

In the case of a std::shared_ptr or a fbl::RefPtr, this causes an AddRef/Remove
cycle which is not needed.  In the case of the fbl::RefPtr tests, which are
designed to make sure that AddRef/Release only happens exactly when expected,
this causes the tests to fail.

For now, work around this issue by std::decay'ing the type passed to the
ASSERT_VAR_COERCE macro, and then static casting to a const& version of that
type instead.  In the long run, a better solution might be to attempt to avoid
any sort of coercion in these macros/lambdas at all.

Change-Id: Idddddb80b9a6df24951abaa7d8b63e8908c905cf