[debugger] List previously connected processes.
The client now asks the agent for any processes it is already connected
to and sends a notification to the console.

Currently there is nothing the user can do. This is not really an issue
as people don't (normally) re-connect to an existing instance of the
agent, so this case is pretty much never met.

TEST=Added unit test.

Change-Id: I7216e205f6df014bc345f31fdeb5d176579d094f