[fidlc] Support MAX bounds and fix constant resolution bug
This CL implements the plan laid out in go/i2i-max-bounds for allowing
string:MAX and vector<T>:MAX to indicate values that are bound only by
extrinsic factors. MAX is a constant with value 2^32-1 that can only be
used in the bounds position. As with the built-in type aliases "byte"
and "bytes", it is possible (but not recommended) to shadow MAX with a
local definition.

In addition, this CL:

- Fixes crashes and assertion failures from fxb/37800.
- Removes the unused member variable constants_ in class Library. This
  has been unused since I3ef45afe9100ebd549f76d096987346f8f8d26a2.
- Adds documentation to the similarly-named classes Const, Constant, and
  ConstantValue to explain their different purposes.
- Refactors some repeated code into a new function,
  TypeConstructor::CreateSizeType.

Test: fx build host_x64/fidlc && $FUCHSIA_DIR/out/default.zircon/host-x64-linux-clang/obj/system/utest/fidl-compiler/fidl-compiler-test.debug --case consts_tests

Bug: 37281
Bug: 37800
Change-Id: Ia2fad02b4f61a5965f1a889c32abb2b2e69604c2