[gdc] Remove GDC proxying
Its not very easy to have callbacks in banjo across devhosts.
Due to that reason we decided that we will be having GDC driver in the
same devhost as the camera controller so we no longer need proxying

Since we removed proxying, we needed to convert GDC and GE2D
drivers to composite drivers and add them to the same devhost as the
other drivers.

Test: Ran platform-bus-test, device-enumeration-test

Change-Id: I77882b1d9a9cba3e72b811d947414266ed39eb07