[debugger] Subclass ProcessBreakpoint.
Before a process breakpoint would hold both a SW and a HW breakpoint,
given that they both are very similar in concept. But with the
introduction of Safe Stepping (tm), the behaviour of stepping over a SW
and a HW breakpoint are very different and tracking them both was
getting very cumbersome.

This makes all breakpoints appear as a ProcessBreakpoint and the step
over queueing by the process will take care of correctly running the
differences.

This change only ports over the SW breakpoints, so HW breakpoints are
disabled until ported. As there are no real users of HW breakpoints,
this is fine for now.

TEST=Ported the test to Software Breakpoint.
     Manually tested breakpoints on both /boot/bin/ps and the debug
     agent.

Change-Id: I7191cfad037d816fdd566b629502fce3a3ff4d1d