[debugger] Debug Agent lists processes on limbo.
It asks this by querying a LimboProvider object, that it's in charge of
obtaining this information.

Currently the normal functionality is not hooked up, as there will be
no LimboProvider present (right now it's only set on tests).

When further tests if provided (actually calling the FIDL protocol),
this can be enabled.

Extended the MockObjectProvider to provide more ergonomic ways of
querying the data.

TEST=Added unit test.

Change-Id: I7943092d1a614a98b7891828d48fb11e329d42f9