[debugger] Expand symbol names for all namespaces.
Adds a "permissive" expansion helper for InputLocations that expands
symbol names based on a language-aware namespace traversal to absolute
names that the symbol system can return information about.

Use this new facility for resolving bookmarks. It is not hooked up to
other commands, this will be done in a followup.

Change how anonymous names are handled. Previuosly anonymous namespaces
were named "(anon)" by the namespace code that generates the identifier.
This means the internal name of any symbol inside it has that weird
"(anon)" string in it.

With this change, the namespace now inserts an empty name and converting
to a user-visible string is done at the presentation later. This makes
things match the index and we have more control over presentation. It
changes the string to "$anon" to match the current escaping proposal,
but input is still not handled so these symbols are still impossible to
type.

The breakpoint tests had to be updated to provide symbol index
information so the symbols are found using the new system.

Adds some additional debugging capabilities to the Locatino object.

BUG=37405

Change-Id: I1102ce754e8833894f50b0c5ad6a871f2bfb2f8b