[upload_debug_symbols] Reland "Add per-object upload timeout"
This reverts commit 1e21be045b160a34c9d2adf433408dfcb13de1b0.

Reason for revert: Have fix for race condition. Previously
it was possible for the goroutine to write to a closed
channel if the context was closed first.

This is reproducible by running go test --race. After this
change, a race warning is no longer seen. The original
test demonstrates the change in behavior as it fails before
the change and succeeds after.

Original: https://fuchsia-review.googlesource.com/c/fuchsia/+/325567
Diff: https://fuchsia-review.googlesource.com/c/fuchsia/+/326509/1..4

Bug: 38082

Change-Id: I9e46577fd5071f7f285cdfe612e2e7e315419da2