[kazoo][abigen] Re-add some argument attributes to json output
I screwed this up in
https://fuchsia-review.googlesource.com/c/fuchsia/+/313366/. I thought
nothing needed the argument attributes, but it turns out the
documentation generator needs "IN" to properly specify const. So, readd
those to abigen, and add them to kazoo. This makes a run of
zircon/scripts/update-docs-from-abigen be a no-op (where currently it
will delete a bunch of "const"s from the docs.)

Bug: 33668
Change-Id: I1184c7f220b102daf861f8be7c2be9a7c5c58178