[audio_core] Prime resamplers before assessing audio fidelity
For resamplers that use a significant amount of "history", the way we
currently measure audio fidelity puts them at a disadvantage, because
we essentially do not evaluate fidelity in a "steady-state" mode,
but rather start evaluating their immediate output (including any
initial transients). In reality this is not a significant penalty for
the LinearInterpolator resampler (and not an issue at all for the
SampleAndHold resampler), but this will be a problem for upcoming
resamplers that use convolution filters of considerable width.

With this change, our fidelity tests "prime" each resampler with the
final frames of the test buffer (these buffers are periodic, so final
frames are equivalent to what would have immediately preceded the
first frames of the buffer). Also, the DCHECK that required the
starting source position to be strictly less than the end of the
source buffer has been slightly relaxed, so that source position can
also be exactly equal to the end of the source buffer. This is
understood to be a priming operation: no output frame is produced,
TRUE is returned (indicating that the source buffer has been
completely consumed), and the final frames of the buffer are retained
in the resampler's "history cache" (if it uses one).

Test: only behavior change for existing product code is the relaxation
of a DCHECK that is not triggered by any automated test, tool or user
interaction. There are test code changes to first prime the samplers
before assessing audio fidelity.

Change-Id: Ia3638af30e2abe45302bfe3b1314851f9820c9d2