[FDF] Disable flaky test
The test started flaking back but a different place this time.
The dispatcher itself seemed to be invalid when the async_cancel_wait is
called and resulted in a hard crash. This seems to be test problem to
me. Need sometime to repro and analyze.
Disable the test until then to unblock CQ.

Test: fx build. fx run-test coordinator-test.

Change-Id: I4caea59c135a1d6ed80768a9eb1b431cdce809a1