[cmgr] Timer-based WorkScheduler
This change implements a timer for disptaching work via WorkScheduler.
The timer is re-evaluated when the collection of scheduled work items or
the dispatch period might have changed.

Presently, dispatching work via a FIDL interface to components is not
implemented. When WorkUnit instances are dispatched, they are simply
removed from the collection of scheduled work items in WorkScheduler.

Test: fx shell run-test-component component_manager_tests.cmx --nocapture 'work_scheduler_time'
Bug: 35214
Bug: 35216

Change-Id: I38b5b44d12c96e1dd835821e66222348b3d86fd2