[ledger][benchmark] Track new DB creations instead of cached DB creations.
add_new_page(_precached).tspec benchmarks would track the duration of cached DB
creation. When running these benchmarks we have generally observed that the for
each DB created, there is one cached DB that is created before.

However, if this benchmark runs very fast on a device, it can happen that the
number of cached DBs created is smaller than the total and thus, the expected
number of traces cannot accurately be predicted. This has actually been observed,
causing a flake in issue 37357.

This CL updates the tspec files to measure the time to create a DB, whether it
is a cached or not. This way the expected number can accuretly be predicted,
idependently of how fast the benchmark runs.

BUG: 37357
Change-Id: I3ff54cbc838151790146fddda2bb98a116c07c52