[media.sessions] Implement Filter for events
Player events come from each player. Each client may have a distinct set
of filters on those events. The filters depend on the state of the
player at the time of the event.

Currently, this is resolved by having each client's dispatcher cache
some kind of state for the player. This is complicated and indirect.

This filter will allow the source of the events, which keeps all player
state, to emit events already tagged with what filters they should be
allowed to pass through, and client dispatchers can blindly apply their
filter and emit the events that satisfy it.

TEST: fx run-test mediasession_tests
Change-Id: I1aa448908c062d80ef83b932a6ea292efc3779b3