[debugger] Index cross-unit references.
Handles DW_FORM_ref_addr references when indexing. These can be seen
sometimes in optimized code in the DW_AT_specification attribute where
the declaration is in a separate unit that the code.

The normal optimized index pass doesn't work for these references since
it assumes everything is in the same unit. A new code path is added that
uses LLVM's getParent() call in O(log n) time instead of our O(1)
in-unit lookup table. Since a minority of functions use this form, the
performance hit is very low.

Since we have not produced a small DWARF file with this attribute, the
new codepath is tested by forcing the slow path for the existing tests.
This ensures that the slow and fast paths both produce the same results.

BUG=37726

Change-Id: I6c8d790237f8a2a1a78637622ceb0c8e4d4dbec5