[escher] Remove external references to command buffer impl
This is to clean up Escher code which was littered with exposed
references to the command buffer's underlying implementation class.
Now all calls to the command buffer are direct, which will make it
easier to deprecate the impl class in the future.

Testing: This is a pure refactor, so no additional unit tests
are needed.

Change-Id: I7945d3981457d1571692f03da64933c03f6ad60d