[audio_core] Filter class
This CL adds another small class to AudioCore, preparing for the
upcoming high-quality audio resampler. The Filter class implements
a convolution-based filter, based on the rate ratio and the number
of taps. There are three subclasses: PointFilter, LinearFilter and
SincFilter.  Only the latter will be immediately used, as this filter
includes the required low-pass filter (for down-sampling). However,
the Point and Linear specializations are useful for proving out the
Filter class implementation.

The next CL will assemble all the pieces (PositionManager,
ChannelStrip, SincFilter) to introduce the SincSampler class.

Test: 13 unittests added to audio_mixer_unittests.

Change-Id: I50991adb53c0ca57f217360a1fc771727e33bdf5