[activity] Make out-of-order event handling per client
Currently, if a Tracker client sends an activity that is before the
system-wide last transition time, the client will receive a
ZX_ERR_OUT_OF_RANGE and the channel will terminate. (Whether the channel
*should* terminate is an open question, but this is the current state of
the world).

This prevents clients from sending out-of-order activities (which is
desirable), but can lead to raciness between competing clients that can
lead to random occurrences of ZX_ERR_OUT_OF_RANGE.

Change this to making it so that a per-client last event time is
maintained. This prevents clients from sending out-of-order messages (as
desired), but prevents clients from interfering with one another.

Tested: Unit testing.
Change-Id: Ifccb38e414261c0ae57d7bead9883b1a9343a675