[zircon][host][fvm]: max-disk-size param to host.
Add the parameter to define a preallocated metadata size for generating
the sparse image.

After this is done, all that remains is updating the product definition
to generate the image with the preallocated size.

Because there are no parameters changed, there should be no changes yet
to the CQ. Though now one could manually generate the image and trigger
an OTA test.

TEST=build,CQ

Change-Id: I82a84de6b5627c2a564a08d2c6b0a589873ce85c