[blackout] add unit tests for host library
The tests are for the various test steps the host side of the test
framework can take. For the most part, the behavior of these tests is
straightforward - just run a subcommand on the target device. However,
the verification step has some more complex logic about error conversion
and retry loops that is tested more thoroughly.

Test: fx run-host-tests blackout_host_lib_test
Bug: 34076
Change-Id: Id5fe66321d92127bdab2e78028f2bb44750bab87