[fuzzing][crypto] add hmac-fuzzer
This CL adds hmac-fuzzer for testing the HMAC::Create function by
fuzzing all of its parameters.

HMAC::Create takes the following arguments: a digest algorithm, a key,
the payload and its size, some flags and a pointer in which to store the
result. If the function succeeds, it will return ZX_OK.

The fuzzer makes use of `FuzzedDataProvider` to pick up the parameters
for HMAC::Create. The digest algorithm is either kSHA256 or
kUninitialized. There is code to short-circuit the flags, so the fuzzer
picks a flag that is a combination of ALLOW_TRUNCATION and
ALLOW_WEAK_KEYS, but also sometimes picks a completely random flag.

The key that will be used in hmac is copied from the fuzz buffer, capped
at 1024 bytes.

HMAC::Create is then called to compute the hmac of the remaining data in
the fuzz buffer.

After the hmac is computed, if it succeeds, the HMAC::Verify function
will be called, and it is expected for it to succeed.

Change-Id: I0fc9a7f0f7932410169b54c9223db5be4fa17ea9