[kernel][x86] CpuId: Add test data for a Skylake-U (i5-6260U) CPU
CpuId has example data for a number of common CPUs on which we expect
to run Fuchsia; before this commit, it had Intel Broadwell-EP,
Intel Goldmont (Apollo Lake SOC), AMD Threadripper, and AMD Stoney
Ridge (APU).

Add test data for Skylake-U; Skylake/KabyLake-U are the most common
x86 hardware we run Fuchsia on, powering Eve, Dawson Canyon NUCs,
and Tolouse. It is useful to have that data available for unit tests.

Add data from a Skylake-U updated to the most recent MCU, which
allows it to enumerate MD_CLEAR.

Change-Id: Icd6f0ac3797205db392c83482ca17dfd8d7eb686