[bt][l2cap] Revoke queued ACL data packets for closed L2CAP channel
Adds ACLDataChannel::DropQueuedPackets method that is called when an L2CAP
channel is closed to revoke packets in the ACLDataChannel queue that are
for that channel.

When channels are closed, LogicalLink will use DropQueuedPackets to filter
packets for the closed channel.

Bug: 1547

Test: in bt-host-unittests
HCI_ACLDataChannelTest.FilterQueueRemovesPacketsMatchingFilterFromQueue
L2CAP_ChannelManagerTest.ACLOutboundDynamicChannelLocalDisconnect
L2CAP_ChannelManagerTest.ACLOutboundDynamicChannelRemoteDisconnect

Change-Id: I7b12f34fc35c7782fc69b59937ee54cf0073be59