[minfs][test] Provide more lenient error-handling in test setup
The result of "Sync" may or may not be a failure in the steup
functions of the Journal Integration tests. These failures are not
very interesting however, since this is part of test setup to
"force the device into a partially-written state".

Since testing the result of these functions is not the objective of
these tests, provide more lenient error-handling.

Bug: 38165

Change-Id: I5d8e8afea0690c5c0e9decc3aea91aca321a9e89