[workstation] remove userpicker_base_shell
Removes the references to userpicker_base_shell since
it is a direct dependency of ermine.
This CL requires fxr/326639 to land before it can
be merged to avoid breaking the workstation product.

Base shells will no longer be part of the session framework
so ermine will directly implement the login screen as we
migrate to the session framework. Having the dependency be declared
by ermine instead of the workstation product will allow ermine
to delete the component without needing to update the workstation
product configuration as well.

Change-Id: I92d2a37a8031005deb01d88819a774f38cfd45a3