[virtcon] Convert to use sysmem
In the future we want to remove the legacy ComputeLinearImageStride and
AllocateVmo calls on the display controller, since they're redundant
with sysmem's memory allocation system (and not flexible enough to work
with things like vulkan). virtcon is one of the last pieces of code
using those calls, so switch it over to using sysmem.

Test: vc-display-test
Bug: 33157

Change-Id: I138bc0cdbfd4ed33b8d5aab41db3054d00de154b