[display] Add capture type for Image buffer types
Display drivers need a way to distinguish between a buffer meant for
reading out an image (in normal display cases) and a buffer meant for writing back an image (in
case of capture). The type of buffer is needed when setting the
constraints for buffer collection. This CL adds a new type for
capture-type of images.

Also removed display-controller.h header and updated code to use the
banjo header

Test: Compiles. Not change in behavior. Ran display-test to ensure the
renaming of layer_t to customer_layer_t doesn't cause issue.

Change-Id: Ibf70b083d25f24761cfccda3650b33ae11972b36