[debugger] Output more info in index dump.
The current index dump can't distinguish between finding a function
declaration and a function implementation, and whether multiple function
implementations have been found.

Adds support for optionally dumping this information. This is used in
the dump test.

The TestSymbolModule is changed to remove the dedicated loading logic
and to use a regular ModuleSymbolsImpl. Indexing is made optional to
support this test case.

The unit pointer on all LazySymbols is removed. This was duplicitave
with the offset but I didn't understand what offsets were when I wrote
it.

There should be no user-observable behavior change.

Change-Id: I1cda8d317da17b308683498f42ed4f3399ae6788