[go][inspect] Fix test failure
The TestReverseFree test was broken, causing flakes picked up in the
referenced bug. Specifically, the full tree was not expressed when
matching the blocks, which caused the goroutine that iterated over the
heap to continue after finding the first block (as it was not the final
one). Often, this goroutine would finish before the rest of the test,
but when it did not, the test function would return, closing the heap.
This would result in the backing VMO disappearing, and when the
goroutine continued, it'd fail trying to read pages that weren't mapped.

This change removes the scanBlocks function entirely, moving its
contents into the matchLayout function. The asynchronous nature made
this a bit annoying to debug, which is why that was done.

The actual fix includes an addition of matchLayout that returns a fatal
error if the entire tree isn't expressed, as well as expressing the
entire tree in TestReverseFree.

Bug: 37325
Change-Id: Icbe5fb27fd8d03bed597d8c9b23be27c0bec1fe4