[Setting] Reland Volume Control API
This CL relands
https://fuchsia-review.googlesource.com/c/fuchsia/+/325417

The CL got reverted because of potential issues with
connecting the audio core service at start up.

The reland addresses the issue by handling the failure
to connect to the audio core service. Audio Controller
will store a flag that indicates if it manages to connect
to the service. If it fails, it will try again when it
receives a Get or SetVolume request.

StreamVolumeControl is also modified to handle the failure
if it fails to bind a VolumeControl proxy to the audio
core service. StreamVolumeControl now stores the
VolumeControl as an Option and if the proxy is None, then
it'll try to create and bind a new proxy when it receives
a request to set the volume.

Bug: fxb/37403
Testing:
There are no new behavior changes. There are changes to
audio_test.rs so that it’ll test that the volume is set correctly
through Volume Control instead of Audio Core
You can run the tests via:
   fx run-test setui_service_tests -- --test

Change-Id: Ic712968fc7e742e2b86334a754025978514d2cd0