[blackout] add additional run modes
This adds more run modes to the harness. Originally there was just one
run mode - run the test once. This adds 2 additional modes, configured
at runtime based on command line flags. The modes are -

- run once
- run N iterations
- run N iterations or until failure

In service of these new modes, a new abstraction over seeds is
introduced. A seed can either be constant, when it's set directly by the
command line, in which case it never changes even during the
multiple-test run modes, or it can be variable, in which case it is
re-rolled for every test run in the multi-run modes.

Change-Id: Ic16027faffbeb6e568ba20aaab481f185b56a565